Moved kindergeld to dag.
