+++ gettsim/taxes/zu_versteuerndes_eink.py
@@ -77,9 +77,7 @@ def kinderfreibetrag(tax_unit, params, kindergeld_params):
 
     # Count number of children eligible for Child Benefit.
     # Child allowance is only received for these kids.
-    kigeld_kinder = kindergeld_params["kindergeld_anspruch_regel"](
-        tax_unit, kindergeld_params
-    ).sum()
+    kigeld_kinder = tax_unit["_kindergeld_anspruch"].sum()
 
     # Find out who has the lower zve among partners
     nokfb_lower = tax_unit["_zu_versteuerndes_eink_kein_kind_freib"].min()
