+++ gettsim/taxes/zu_versteuerndes_eink_dag.py
@@ -36,6 +36,7 @@ def _zu_versteuerndes_eink_kein_kind_freib(
     brutto_eink_6,
     brutto_eink_7,
     sum_brutto_eink,
+    _kindergeld_anspruch,
     hh_freib,
     altersfreib,
     behinderungsgrad_pauschalbetrag,
@@ -76,6 +77,7 @@ def _zu_versteuerndes_eink_kein_kind_freib(
             brutto_eink_6,
             brutto_eink_7,
             sum_brutto_eink,
+            _kindergeld_anspruch,
             altersfreib,
             behinderungsgrad_pauschalbetrag,
             wohnort_ost,
